+++ crates/web-sys/src/features/gen_ServiceWorkerRegistration.rs
@@ -84,17 +84,6 @@ extern "C" {
     pub fn get_notifications(
         this: &ServiceWorkerRegistration,
     ) -> Result<::js_sys::Promise, JsValue>;
-    #[cfg(feature = "GetNotificationOptions")]
-    # [wasm_bindgen (catch , method , structural , js_class = "ServiceWorkerRegistration" , js_name = getNotifications)]
-    #[doc = "The `getNotifications()` method."]
-    #[doc = ""]
-    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/ServiceWorkerRegistration/getNotifications)"]
-    #[doc = ""]
-    #[doc = "*This API requires the following crate features to be activated: `GetNotificationOptions`, `ServiceWorkerRegistration`*"]
-    pub fn get_notifications_with_filter(
-        this: &ServiceWorkerRegistration,
-        filter: &GetNotificationOptions,
-    ) -> Result<::js_sys::Promise, JsValue>;
     # [wasm_bindgen (catch , method , structural , js_class = "ServiceWorkerRegistration" , js_name = showNotification)]
     #[doc = "The `showNotification()` method."]
     #[doc = ""]