+++ CHANGELOG.md
@@ -67,6 +67,9 @@
   proc-macro and accept the `--include-ignored` flag with `wasm-bindgen-test-runner`.
   [#3644](https://github.com/rustwasm/wasm-bindgen/pull/3644)
 
+* Added missing additions to the Notification API.
+  [#3667](https://github.com/rustwasm/wasm-bindgen/pull/3667)
+
 ### Changed
 
 * Updated the WebGPU WebIDL.
@@ -156,6 +159,9 @@
 
 * Removed `ReadableStreamByobReader::read_with_u8_array()` because it doesn't work with Wasm.
   [#3582](https://github.com/rustwasm/wasm-bindgen/pull/3582)
+
+* Removed `GetNotificationOptions`, `NotificationBehavior` and `Notification.get()` because
+  they don't exist anymore.
 
 ## [0.2.87](https://github.com/rustwasm/wasm-bindgen/compare/0.2.86...0.2.87)